Use UUID instances for compatibility with uuid changes in https://github.com/libkeepass/pykeepass/commit/616a23f7889a1c5274c2479bc4c714b1fb084c3b Move BaseElement.deref() and BaseElement.ref() to Entry, field references aren't supported for groups in KeePass.
